@@ -1,7 +1,7 @@
 name: connectivity_for_web
 description: An implementation for the web platform of the Flutter `connectivity` plugin. This uses the NetworkInformation Web API, with a fallback to Navigator.onLine.
-version: 0.3.1+4
 repository: https://github.com/flutter/plugins/tree/master/packages/connectivity/connectivity_for_web
+version: 0.4.0
 
 flutter:
   plugin:
@@ -11,21 +11,16 @@ flutter:
         fileName: connectivity_for_web.dart
 
 dependencies:
-  connectivity_platform_interface: ^1.0.3
+  connectivity_platform_interface: ^2.0.0
   flutter_web_plugins:
     sdk: flutter
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  flutter_driver:
-    sdk: flutter
   flutter_test:
     sdk: flutter
-  integration_test:
-    path: ../../integration_test
-  mockito: ^4.1.1
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.4"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.20.0"
