@@ -1,7 +1,7 @@
 name: google_maps_flutter_web
 description: Web platform implementation of google_maps_flutter
 homepage: https://github.com/flutter/plugins/tree/master/packages/google_maps_flutter
-version: 0.1.2
+version: 0.2.1
 
 flutter:
   plugin:
@@ -15,22 +15,17 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  meta: ^1.1.7
-  google_maps_flutter_platform_interface: ^1.1.0
+  meta: ^1.3.0
+  google_maps_flutter_platform_interface: ^2.0.1
   google_maps: ^3.4.5
-  stream_transform: ^1.2.0
-  sanitize_html: ^1.4.1
+  stream_transform: ^2.0.0
+  sanitize_html: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  http: ^0.12.2
-  url_launcher: ^5.2.5
-  pedantic: ^1.8.0
-  mockito: ^4.1.1
-  integration_test:
-    path: ../../integration_test
+  pedantic: ^1.10.0
 
 environment:
   sdk: ">=2.3.0 <3.0.0"
-  flutter: ">=1.10.0"
+  flutter: ">=1.20.0"
