@@ -8,7 +8,6 @@ import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:flutter_plugin_tools/src/xctest_command.dart';
 import 'package:test/test.dart';
-import 'package:flutter_plugin_tools/src/common.dart';
 
 import 'mocks.dart';
 import 'util.dart';
@@ -81,7 +80,6 @@ final _kDeviceListMap = {
 
 void main() {
   const String _kDestination = '--ios-destination';
-  const String _kTarget = '--target';
   const String _kSkip = '--skip';
 
   group('test xctest_command', () {
@@ -98,12 +96,6 @@ void main() {
       runner = CommandRunner<Null>('xctest_command', 'Test for xctest_command');
       runner.addCommand(command);
       cleanupPackages();
-    });
-
-    test('Not specifying --target throws', () async {
-      await expectLater(
-          () => runner.run(<String>['xctest', _kDestination, 'a_destination']),
-          throwsA(const TypeMatcher<ToolExit>()));
     });
 
     test('skip if ios is not supported', () async {
@@ -123,8 +115,6 @@ void main() {
       processRunner.processToReturn = mockProcess;
       final List<String> output = await runCapturingPrint(runner, <String>[
         'xctest',
-        _kTarget,
-        'foo_scheme',
         _kDestination,
         'foo_destination'
       ]);
@@ -134,106 +124,7 @@ void main() {
       cleanupPackages();
     });
 
-    test('running with correct scheme and destination, did not find scheme',
-        () async {
-      createFakePlugin('plugin',
-          withExtraFiles: <List<String>>[
-            <String>['example', 'test'],
-          ],
-          isIosPlugin: true);
-
-      final Directory pluginExampleDirectory =
-          mockPackagesDir.childDirectory('plugin').childDirectory('example');
-
-      createFakePubspec(pluginExampleDirectory, isFlutter: true);
-
-      final MockProcess mockProcess = MockProcess();
-      mockProcess.exitCodeCompleter.complete(0);
-      processRunner.processToReturn = mockProcess;
-      processRunner.resultStdout = '{"project":{"targets":["bar_scheme"]}}';
-
-      await expectLater(() async {
-        final List<String> output = await runCapturingPrint(runner, <String>[
-          'xctest',
-          _kTarget,
-          'foo_scheme',
-          _kDestination,
-          'foo_destination'
-        ]);
-        expect(output,
-            contains('foo_scheme not configured for plugin, test failed.'));
-        expect(
-            processRunner.recordedCalls,
-            orderedEquals(<ProcessCall>[
-              ProcessCall('xcrun', <String>['simctl', 'list', '--json'], null),
-              ProcessCall(
-                  'xcodebuild',
-                  <String>[
-                    '-project',
-                    'ios/Runner.xcodeproj',
-                    '-list',
-                    '-json'
-                  ],
-                  pluginExampleDirectory.path),
-            ]));
-      }, throwsA(const TypeMatcher<ToolExit>()));
-      cleanupPackages();
-    });
-
-    test('running with correct scheme and destination, found scheme', () async {
-      createFakePlugin('plugin',
-          withExtraFiles: <List<String>>[
-            <String>['example', 'test'],
-          ],
-          isIosPlugin: true);
-
-      final Directory pluginExampleDirectory =
-          mockPackagesDir.childDirectory('plugin').childDirectory('example');
-
-      createFakePubspec(pluginExampleDirectory, isFlutter: true);
-
-      final MockProcess mockProcess = MockProcess();
-      mockProcess.exitCodeCompleter.complete(0);
-      processRunner.processToReturn = mockProcess;
-      processRunner.resultStdout =
-          '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
-      List<String> output = await runCapturingPrint(runner, <String>[
-        'xctest',
-        _kTarget,
-        'foo_scheme',
-        _kDestination,
-        'foo_destination'
-      ]);
-
-      expect(output, contains('Successfully ran xctest for plugin'));
-
-      expect(
-          processRunner.recordedCalls,
-          orderedEquals(<ProcessCall>[
-            ProcessCall(
-                'xcodebuild',
-                <String>['-project', 'ios/Runner.xcodeproj', '-list', '-json'],
-                pluginExampleDirectory.path),
-            ProcessCall(
-                'xcodebuild',
-                <String>[
-                  'test',
-                  '-workspace',
-                  'ios/Runner.xcworkspace',
-                  '-scheme',
-                  'foo_scheme',
-                  '-destination',
-                  'foo_destination',
-                  'CODE_SIGN_IDENTITY=""',
-                  'CODE_SIGNING_REQUIRED=NO'
-                ],
-                pluginExampleDirectory.path),
-          ]));
-
-      cleanupPackages();
-    });
-
-    test('running with correct scheme and destination, skip 1 plugin',
+    test('running with correct destination, skip 1 plugin',
         () async {
       createFakePlugin('plugin1',
           withExtraFiles: <List<String>>[
@@ -260,8 +151,6 @@ void main() {
           '{"project":{"targets":["bar_scheme", "foo_scheme"]}}';
       List<String> output = await runCapturingPrint(runner, <String>[
         'xctest',
-        _kTarget,
-        'foo_scheme',
         _kDestination,
         'foo_destination',
         _kSkip,
@@ -275,21 +164,22 @@ void main() {
           processRunner.recordedCalls,
           orderedEquals(<ProcessCall>[
             ProcessCall(
-                'xcodebuild',
-                <String>['-project', 'ios/Runner.xcodeproj', '-list', '-json'],
-                pluginExampleDirectory2.path),
-            ProcessCall(
-                'xcodebuild',
+                'xcrun',
                 <String>[
+                  'xcodebuild',
                   'test',
+                  'analyze',
                   '-workspace',
                   'ios/Runner.xcworkspace',
+                  '-configuration',
+                  'Debug',
                   '-scheme',
-                  'foo_scheme',
+                  'Runner',
                   '-destination',
                   'foo_destination',
                   'CODE_SIGN_IDENTITY=""',
-                  'CODE_SIGNING_REQUIRED=NO'
+                  'CODE_SIGNING_REQUIRED=NO',
+                  'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                 ],
                 pluginExampleDirectory2.path),
           ]));
@@ -324,8 +214,6 @@ void main() {
           jsonEncode(schemeCommandResult..addAll(_kDeviceListMap));
       await runner.run(<String>[
         'xctest',
-        _kTarget,
-        'foo_scheme',
       ]);
 
       expect(
@@ -333,21 +221,22 @@ void main() {
           orderedEquals(<ProcessCall>[
             ProcessCall('xcrun', <String>['simctl', 'list', '--json'], null),
             ProcessCall(
-                'xcodebuild',
-                <String>['-project', 'ios/Runner.xcodeproj', '-list', '-json'],
-                pluginExampleDirectory.path),
-            ProcessCall(
-                'xcodebuild',
+                'xcrun',
                 <String>[
+                  'xcodebuild',
                   'test',
+                  'analyze',
                   '-workspace',
                   'ios/Runner.xcworkspace',
+                  '-configuration',
+                  'Debug',
                   '-scheme',
-                  'foo_scheme',
+                  'Runner',
                   '-destination',
                   'id=1E76A0FD-38AC-4537-A989-EA639D7D012A',
                   'CODE_SIGN_IDENTITY=""',
-                  'CODE_SIGNING_REQUIRED=NO'
+                  'CODE_SIGNING_REQUIRED=NO',
+                  'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                 ],
                 pluginExampleDirectory.path),
           ]));
